refactor: Movement 입력 검증을 Validator에서 BridgeGame으로 옮김
- 검증에 대한 책임을 BridgeGame이 지도록 함
